refactored "Annee", "Departement" and "Formation" classes + Updated getters in frontend

+++ projet-backend/Projet/Projet/src/main/java/com/fst/il/m2/Projet/repositories/ResponsableDepartementRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResponsableDepartementRepository extends JpaRepository<ResponsableDepartement, Integer> {
+public interface ResponsableDepartementRepository extends JpaRepository<ResponsableDepartement, Long> {
     @Modifying
     @Query("DELETE FROM ResponsableDepartement r WHERE r.user = :user")
     void deleteByUser(@Param("user") User user);
