+++ projet-backend/Projet/Projet/src/main/java/com/fst/il/m2/Projet/repositories/ResponsableFormationRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface ResponsableFormationRepository extends JpaRepository<ResponsableFormation, Integer> {
+public interface ResponsableFormationRepository extends JpaRepository<ResponsableFormation, Long> {
     @Modifying
     @Query("DELETE FROM ResponsableFormation r WHERE r.user = :user")
     void deleteByUser(@Param("user") User user);
